requirements.txt
crewai
langchain
langchain-community
langchain-core
ollama
python-dotenv
pydantic 